updated to naviisml/libft@latest

+++ srcs/flags/parse_number.c
@@ -6,7 +6,7 @@
 /*   By: navi <navi@student.codam.nl>                 +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2021/11/17 14:03:25 by navi          #+#    #+#                 */
-/*   Updated: 2021/11/19 17:12:05 by nismail       ########   odam.nl         */
+/*   Updated: 2021/12/07 02:41:26 by navi          ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ static int	ft_putuint_fd(unsigned int nbr, int fd)
 {
 	int	len;
 
-	len = ft_floor(nbr);
+	len = ft_floor((__int128)nbr);
 	if (nbr > 9)
 	{
 		ft_putuint_fd((nbr / 10), fd);
@@ -48,7 +48,7 @@ int	util_parse_uint(t_print *arg, char c)
 	int				len;
 
 	nbr = va_arg(arg->args, unsigned int);
-	len = ft_floor(nbr);
+	len = ft_floor((__int128)nbr);
 	util_print_prefix(arg, (long long)nbr);
 	ft_putuint_fd(nbr, arg->fd);
 	arg->len += len;
@@ -68,7 +68,7 @@ int	util_parse_int(t_print *arg, char c)
 	int	len;
 
 	nbr = va_arg(arg->args, int);
-	len = ft_floor(nbr);
+	len = ft_floor((__int128)nbr);
 	util_print_prefix(arg, (long long)nbr);
 	ft_putnbr_fd(nbr, arg->fd);
 	arg->len += len;
